[#1] refactor: Apply queryDSL
