--- BE/greenEarth/src/main/java/com/ssafy/greenEarth/repository/GreenEarthLogRepository.java
@@ -1,8 +1,0 @@
-package com.ssafy.greenEarth.repository;
-
-import com.ssafy.greenEarth.domain.GreenEarthLog;
-import com.ssafy.greenEarth.domain.GreenEarthLogId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GreenEarthLogRepository extends JpaRepository<GreenEarthLog, GreenEarthLogId> {
-}
