--- BE/greenEarth/src/main/java/com/ssafy/greenEarth/repository/RewardRepository.java
@@ -1,12 +1,0 @@
-package com.ssafy.greenEarth.repository;
-
-import com.ssafy.greenEarth.domain.Reward;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RewardRepository extends JpaRepository<Reward, Integer> {
-
-    Optional<Reward> findRewardById(int id);
-
-}
