+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/service/AuthService.java
@@ -1,19 +1,25 @@
 package com.ssafy.greenEarth.service;
 
-import com.ssafy.greenEarth.domain.*;
-import com.ssafy.greenEarth.dto.Auth.*;
+import com.ssafy.greenEarth.domain.Child;
+import com.ssafy.greenEarth.domain.RefreshToken;
+import com.ssafy.greenEarth.domain.RefreshTokenId;
+import com.ssafy.greenEarth.domain.Role;
+import com.ssafy.greenEarth.dto.Auth.LoginDto;
+import com.ssafy.greenEarth.dto.Auth.TokenDto;
+import com.ssafy.greenEarth.exception.BusinessException;
 import com.ssafy.greenEarth.jwt.TokenProvider;
-import com.ssafy.greenEarth.repository.*;
-
+import com.ssafy.greenEarth.repository.childRepo.ChildRepository;
+import com.ssafy.greenEarth.repository.parentRepo.ParentRepository;
+import com.ssafy.greenEarth.repository.refreshTokenRepo.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.HashMap;
 import java.util.Map;
 
-import com.ssafy.greenEarth.exception.BusinessException;
 import static com.ssafy.greenEarth.exception.ErrorCode.*;
 
 @Slf4j
