+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/service/GameService.java
@@ -6,9 +6,9 @@ import com.ssafy.greenEarth.domain.GreenEarthLogId;
 import com.ssafy.greenEarth.dto.Game.MileageAddReqDto;
 import com.ssafy.greenEarth.dto.Member.ChildProfileDto;
 import com.ssafy.greenEarth.exception.BusinessException;
-import com.ssafy.greenEarth.repository.ChildRepository;
-import com.ssafy.greenEarth.repository.GreenEarthLogRepository;
-import com.ssafy.greenEarth.repository.GreenEarthRepository;
+import com.ssafy.greenEarth.repository.childRepo.ChildRepository;
+import com.ssafy.greenEarth.repository.greenEarthLogRepo.GreenEarthLogRepository;
+import com.ssafy.greenEarth.repository.greenEarthRepo.GreenEarthRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
