+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/service/MissionService.java
@@ -6,7 +6,12 @@ import com.ssafy.greenEarth.dto.Mission.MissionPutDto;
 import com.ssafy.greenEarth.dto.Mission.MissionReqDto;
 import com.ssafy.greenEarth.dto.Mission.MissionResDto;
 import com.ssafy.greenEarth.exception.BusinessException;
-import com.ssafy.greenEarth.repository.*;
+import com.ssafy.greenEarth.repository.childRepo.ChildRepository;
+import com.ssafy.greenEarth.repository.greenEarthLogRepo.GreenEarthLogRepository;
+import com.ssafy.greenEarth.repository.greenEarthRepo.GreenEarthRepository;
+import com.ssafy.greenEarth.repository.missionLogRepo.MissionLogRepository;
+import com.ssafy.greenEarth.repository.missionRepo.MissionRepository;
+import com.ssafy.greenEarth.repository.parentRepo.ParentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
