+++ BE/greenEarth/src/main/java/com/ssafy/greenEarth/service/RewardService.java
@@ -4,8 +4,8 @@ import com.ssafy.greenEarth.domain.Child;
 import com.ssafy.greenEarth.domain.Reward;
 import com.ssafy.greenEarth.dto.Reward.*;
 import com.ssafy.greenEarth.exception.BusinessException;
-import com.ssafy.greenEarth.repository.ChildRepository;
-import com.ssafy.greenEarth.repository.RewardRepository;
+import com.ssafy.greenEarth.repository.childRepo.ChildRepository;
+import com.ssafy.greenEarth.repository.rewardRepo.RewardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
